DEEP-CLI/pyproject.toml
[project]
name = "deepseek-cli"
version = "2.0.0"
description = "A powerful AI-powered CLI for DeepSeek models with advanced features"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "DEEP-CLI Team", email = "support@deepcli.com"}
]
keywords = ["deepseek", "ai", "cli", "assistant", "coding"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "openai>=1.0.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.24.0",
    "aiohttp>=3.8.0",
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "toml>=0.10.2",
    "PyYAML>=6.0",
    "pygithub>=2.0.0",
    "aiosqlite>=0.19.0",
    "prompt-toolkit>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
deepcli = "deepcli.main:cli"
deep-cli = "deepcli.main:cli"

[project.urls]
Homepage = "https://github.com/yourusername/deep-cli"
Documentation = "https://deep-cli.readthedocs.io"
Repository = "https://github.com/yourusername/deep-cli"
Issues = "https://github.com/yourusername/deep-cli/issues"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["deepcli*"]
exclude = ["tests*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "C90", "I", "N", "UP", "B", "A", "C4", "SIM", "RUF"]
ignore = ["E501", "B008", "RUF012"]
target-version = "py38"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "--verbose",
    "--cov=deepcli",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--asyncio-mode=auto",
]

[tool.coverage.run]
source = ["src/deepcli"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
]